Move hardware key validation into core, remove obsolete secret-display state
