+++ src/components/secret-display/states.ts
@@ -2,7 +2,6 @@ export const SecretDisplayStates = {
   loading: 'loading',
   revealed: 'revealed',
   noSeed: 'noSeed',
-  securedWithBiometrics: 'securedWithBiometrics',
 } as const;
 
 export type SecretDisplayStatesType = (typeof SecretDisplayStates)[keyof typeof SecretDisplayStates];
