+++ src/features/polymarket/utils/loadViemWallet.ts
@@ -6,8 +6,9 @@ import { LedgerSigner } from '@/features/hardware-wallet/utils/LedgerSigner';
 import * as kc from '@/features/local-auth/keychain';
 import { useBackendNetworksStore } from '@/features/network/stores/backendNetworksStore';
 import { ChainId } from '@/features/network/types/backendNetworks';
+import { isHardwareWalletKey } from '@/features/wallet/core/hardwareWalletKey';
 import { getHdPath, WalletLibraryType } from '@/features/wallet/core/walletLibrary';
-import { isHardwareWalletKey, loadPrivateKey } from '@/features/wallet/data/walletKeychain';
+import { loadPrivateKey } from '@/features/wallet/data/walletKeychain';
 import walletTypes from '@/helpers/walletTypes';
 import { logger, RainbowError } from '@/logger';
 import { getWalletWithAccount } from '@/state/wallets/walletsStore';
