+++ src/features/wallet/data/loadWallet.ts
@@ -4,8 +4,9 @@ import { Wallet } from '@ethersproject/wallet';
 import { type EthereumAddress } from '@/entities/wallet';
 import { LedgerSigner } from '@/features/hardware-wallet/utils/LedgerSigner';
 import { ErrorType } from '@/features/local-auth/keychain';
+import { isHardwareWalletKey } from '@/features/wallet/core/hardwareWalletKey';
 import { getHdPath, WalletLibraryType } from '@/features/wallet/core/walletLibrary';
-import { isHardwareWalletKey, loadAddress, loadPrivateKey } from '@/features/wallet/data/walletKeychain';
+import { loadAddress, loadPrivateKey } from '@/features/wallet/data/walletKeychain';
 import walletTypes from '@/helpers/walletTypes';
 import Navigation from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
